server: deploy the contract [README]
the server is now able to accept a deploy request from the client. however, i've not got the new contract deployed because the address has exausthed its balance on the testnet. additionally, now to use the bnb testnet faucet you're required to have an amount of native cryptocurrency on the mainet as well...)